+++ frontend/src/components/back-button/back-button.html
@@ -1,4 +1,4 @@
-<button aria-label="back icon" class="absolute -left-4 -top-4 bg-secondary rounded-full p-1.5 hover: cursor-pointer" (click)="vm.prev()">
+<button aria-label="back icon" class="absolute -left-4 -top-4 bg-secondary rounded-full p-1.5 hover: cursor-pointer" (click)="prev()">
     <svg xmlns="http://www.w3.org/2000/svg" [attr.height]=height() viewBox="0 -960 960 960" [attr.width]=width() [attr.fill]=fill()>
         <path d="M400-240 160-480l240-240 56 58-142 142h486v80H314l142 142-56 58Z"/>
     </svg>
