+++ frontend/src/components/back-button/back-button.ts
@@ -1,15 +1,18 @@
 import { Component, inject, input } from "@angular/core";
-import { BackButtonViewModel } from "./back-button.vm";
+import { Location } from '@angular/common';
 
 @Component({
-    selector: 'back-button',
-    templateUrl: './back-button.html',
-    providers: [BackButtonViewModel]
+  selector: 'back-button',
+  templateUrl: './back-button.html'
 })
-
 export class BackButton {
-    vm = inject(BackButtonViewModel);
-    height = input<string>("24px");
-    width = input<string>("24px");
-    fill = input<string>('#FFF');
+  private location = inject(Location); // Direkt hier injecten
+  
+  height = input<string>("24px");
+  width = input<string>("24px");
+  fill = input<string>('#FFF');
+
+  prev() {
+    this.location.back(); // Methode direkt in der Komponente
+  }
 }
