+++ frontend/src/services/auth.guard.ts
@@ -1,5 +1,5 @@
 import { CanActivateFn, Router } from "@angular/router";
-import { AuthService } from "./AuthService";
+import { AuthService } from "./auth.service";
 import { inject } from "@angular/core";
 
 export const authGuard: CanActivateFn = (route, state) => {
