+++ frontend/src/services/auth.interceptor.ts
@@ -1,5 +1,5 @@
 import { HttpInterceptorFn } from "@angular/common/http";
-import { AuthService } from "./AuthService";
+import { AuthService } from "./auth.service";
 import { inject } from "@angular/core";
 
 export const authInterceptor: HttpInterceptorFn = (req, next) => {
