+++ frontend/src/services/tour.service.ts
@@ -1,5 +1,5 @@
 import { Injectable, signal, inject } from "@angular/core";
-import { HttpClient } from "@angular/common/http";
+import { HttpClient, HttpParams } from "@angular/common/http";
 import { Tour } from "../app/models/tour.model";
 import { Observable, tap } from "rxjs";
